refactor: 의존성 방향을 고려해서 수정
Order는 Menu 도메인에 의존
