--- src/main/java/kitchenpos/order/application/validator/OrderLineItemsSizeValidator.java
@@ -1,7 +1,0 @@
-package kitchenpos.order.application.validator;
-
-import java.util.List;
-
-public interface OrderLineItemsSizeValidator {
-    void validateOrderLineItems(int orderLineItemsSize, List<Long> menuIds);
-}
